refactor(image-generation): split providers into per-file modules
Convert nanobot/providers/image_generation.py (single 762-line module)
into a package with one module per provider, plus a small registry that
auto-discovers built-ins via pkgutil — mirroring the channel plugin
pattern at nanobot/channels/registry.py.

Adding a new built-in provider is now: drop a file in the package and
re-export the class from __init__.py. The manual register_*() block at
the end of the old module is gone; register_image_gen_provider remains
public for external code that wants to add a provider at runtime.

Tests are split into tests/providers/image_generation/ to mirror the
source package, with shared FakeClient/FakeResponse infrastructure in a
_helpers module. A new test_registry covers discovery, manual
registration, config filtering, and a sanity check that every
discovered provider class is re-exported from the package __init__.

No behavior change. ProvidersConfig and all public import paths are
preserved; existing consumers (the tool, nanobot.py, cli/commands.py,
and external tests) need no edits.
