+++ nanobot/providers/image_generation/openrouter.py
@@ -1,0 +1,111 @@
+"""OpenRouter image generation client."""
+
+from __future__ import annotations
+
+from typing import Any
+
+import httpx
+
+from nanobot.providers.image_generation.base import (
+    GeneratedImageResponse,
+    ImageGenerationError,
+    ImageGenerationProvider,
+    image_path_to_data_url,
+)
+
+_OPENROUTER_ATTRIBUTION_HEADERS = {
+    "HTTP-Referer": "https://github.com/HKUDS/nanobot",
+    "X-OpenRouter-Title": "nanobot",
+    "X-OpenRouter-Categories": "cli-agent,personal-agent",
+}
+
+
+class OpenRouterImageGenerationClient(ImageGenerationProvider):
+    """Small async client for OpenRouter Chat Completions image generation."""
+
+    provider_name = "openrouter"
+    missing_key_message = (
+        "OpenRouter API key is not configured. Set providers.openrouter.apiKey."
+    )
+
+    def _default_base_url(self) -> str:
+        return "https://openrouter.ai/api/v1"
+
+    async def generate(
+        self,
+        *,
+        prompt: str,
+        model: str,
+        reference_images: list[str] | None = None,
+        aspect_ratio: str | None = None,
+        image_size: str | None = None,
+    ) -> GeneratedImageResponse:
+        if not self.api_key:
+            raise ImageGenerationError(self.missing_key_message)
+
+        content: str | list[dict[str, Any]]
+        references = list(reference_images or [])
+        if references:
+            blocks: list[dict[str, Any]] = [{"type": "text", "text": prompt}]
+            blocks.extend(
+                {"type": "image_url", "image_url": {"url": image_path_to_data_url(path)}}
+                for path in references
+            )
+            content = blocks
+        else:
+            content = prompt
+
+        body: dict[str, Any] = {
+            "model": model,
+            "messages": [{"role": "user", "content": content}],
+            "modalities": ["image", "text"],
+            "stream": False,
+        }
+        image_config: dict[str, str] = {}
+        if aspect_ratio:
+            image_config["aspect_ratio"] = aspect_ratio
+        if image_size:
+            image_config["image_size"] = image_size
+        if image_config:
+            body["image_config"] = image_config
+        body.update(self.extra_body)
+
+        headers = {
+            "Authorization": f"Bearer {self.api_key}",
+            "Content-Type": "application/json",
+            **_OPENROUTER_ATTRIBUTION_HEADERS,
+            **self.extra_headers,
+        }
+        url = f"{self.api_base}/chat/completions"
+        response = await self._http_post(url, headers=headers, body=body)
+
+        try:
+            response.raise_for_status()
+        except httpx.HTTPStatusError as exc:
+            detail = response.text[:500]
+            raise ImageGenerationError(f"OpenRouter image generation failed: {detail}") from exc
+
+        data = response.json()
+        images: list[str] = []
+        text_parts: list[str] = []
+        for choice in data.get("choices") or []:
+            if not isinstance(choice, dict):
+                continue
+            message = choice.get("message") or {}
+            if isinstance(message.get("content"), str):
+                text_parts.append(message["content"])
+            for image in message.get("images") or []:
+                if not isinstance(image, dict):
+                    continue
+                image_url = image.get("image_url") or image.get("imageUrl") or {}
+                url_value = image_url.get("url") if isinstance(image_url, dict) else None
+                if isinstance(url_value, str) and url_value.startswith("data:image/"):
+                    images.append(url_value)
+
+        self._require_images(images, data)
+
+        return GeneratedImageResponse(
+            images=images,
+            content="\n".join(part for part in text_parts if part).strip(),
+            raw=data,
+        )
