+++ tests/providers/image_generation/_helpers.py
@@ -1,0 +1,58 @@
+"""Shared HTTP mocks and byte fixtures for image generation provider tests."""
+
+from __future__ import annotations
+
+from typing import Any
+
+import httpx
+
+PNG_BYTES = (
+    b"\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x01"
+    b"\x00\x00\x00\x01\x08\x04\x00\x00\x00\xb5\x1c\x0c\x02"
+    b"\x00\x00\x00\x0bIDATx\xdacd\xfc\xff\x1f\x00\x03\x03"
+    b"\x02\x00\xef\xbf\xa7\xdb\x00\x00\x00\x00IEND\xaeB`\x82"
+)
+PNG_DATA_URL = (
+    "data:image/png;base64,"
+    "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mP8/x8AAwMCAO+/p9sAAAAASUVORK5CYII="
+)
+JPEG_BYTES = b"\xff\xd8\xff\xe0" + b"0" * 12
+RAW_B64 = PNG_DATA_URL.removeprefix("data:image/png;base64,")
+
+
+class FakeResponse:
+    def __init__(
+        self,
+        payload: dict[str, Any],
+        status_code: int = 200,
+        content: bytes = b"",
+    ) -> None:
+        self._payload = payload
+        self.status_code = status_code
+        self.text = str(payload)
+        self.content = content
+        self.request = httpx.Request("POST", "https://openrouter.ai/api/v1/chat/completions")
+
+    def json(self) -> dict[str, Any]:
+        return self._payload
+
+    def raise_for_status(self) -> None:
+        if self.status_code >= 400:
+            response = httpx.Response(self.status_code, request=self.request, text=self.text)
+            raise httpx.HTTPStatusError("failed", request=self.request, response=response)
+
+
+class FakeClient:
+    def __init__(self, response: FakeResponse) -> None:
+        self.response = response
+        self.get_response = response
+        self.calls: list[dict[str, Any]] = []
+        self.get_calls: list[dict[str, Any]] = []
+
+    async def post(self, url: str, **kwargs: Any) -> FakeResponse:
+        self.calls.append({"url": url, **kwargs})
+        return self.response
+
+    async def get(self, url: str, **kwargs: Any) -> FakeResponse:
+        self.get_calls.append({"url": url, **kwargs})
+        return self.get_response
